% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTranscripts.R
\name{cleanTranscripts}
\alias{cleanTranscripts}
\title{Clean raw transcripts}
\usage{
cleanTranscripts(
  transcripts,
  useCellBoundaryFlag = TRUE,
  useSystemControlFlag = TRUE,
  useZstackFlag = TRUE
)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{useCellBoundaryFlag}{a logical to determine if the CellBoundaryFlag column should be used for cleaning the data}

\item{useSystemControlFlag}{a logical to determine if the SystemControlFlag column should be used for cleaning the data}

\item{useZstackFlag}{a logical to determine if the ZstackFlag column should be used for cleaning the data}
}
\value{
A data frame containing only high confidence transcripts.
}
\description{
Function to discard transcripts flagged by setQualityFlags() and setZsliceFlag().
}
\examples{
\dontrun{

cleaned_transcripts <- cleanTranscripts(transcripts)
}
}
