% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StitchRawTranscripts.R
\name{CombineRawTranscripts}
\alias{CombineRawTranscripts}
\title{Combine Raw Transcripts}
\usage{
CombineRawTranscripts(
  in_dir,
  out_dir,
  file_name = "raw_transcripts_stitched.csv"
)
}
\arguments{
\item{in_dir}{A string specifying the path to the input directory containing the raw transcript information for each FOV.}

\item{out_dir}{A string specifying the path to the output directory where the stitched CSV file will be saved.}

\item{file_name}{A string specifying the name of the output CSV file. Defaults to "raw_transcripts.csv".}
}
\value{
A data frame containing the combined raw transcript data from all the files in the input directory.
The function also saves this data frame as a CSV file in the specified output directory.
}
\description{
This function combines the raw transcript data from multiple FOV files in a specified input directory (typically found in AnalysisResults of raw Atomx export data)
and saves the combined data to a CSV file in a specified output directory.
}
\details{
The function reads all the files from the specified input directory. For each subdirectory (representing
a field of view, FOV), it looks for files matching the pattern "complete_code_cell_target_call_coord".
If exactly one such file is found, its data is read and appended to a data frame that accumulates data from all FOVs.
The accumulated data is then written to a CSV file in the specified output directory.

If the input or output directory does not exist, the function stops with an error.
If no matching file or multiple matching files are found in a subdirectory, a warning is issued.
}
\examples{
\dontrun{
raw_transcripts <- CombineRawTranscripts("path/to/input/directory", "path/to/output/directory")
raw_transcripts <- CombineRawTranscripts("path/to/input/directory", "path/to/output/directory", "custom_file_name.csv")
}

}
