% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlags.R
\name{plotFlags}
\alias{plotFlags}
\title{Plot flag summary}
\usage{
plotFlags(transcripts, by = "fov")
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{by}{a character vector specifying the column used to calculate and plot}
}
\value{
A ggplot object.
}
\description{
Initialises a ggplot object to plot the proportion of transcripts flagged by setQualityFlags() and setZsliceFlag().
}
\examples{
\dontrun{

plotFlags(transcripts)
plotFlags(transcripts, by = "SampleID")
}
}
