% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTranscriptByCodeClass.R
\name{PlotTranscriptByCodeClass}
\alias{PlotTranscriptByCodeClass}
\title{Plot Transcripts by Code Class}
\usage{
PlotTranscriptByCodeClass(
  raw_transcripts_file,
  group_by = NULL,
  fig_save = FALSE,
  fig_width = 10,
  fig_height = 8,
  out_dir = NULL
)
}
\arguments{
\item{raw_transcripts_file}{A character string representing the path to a CSV file containing raw transcript data, or a data frame with the transcript data.}

\item{group_by}{A string specifying the column name by which to group the transcript counts along with the code class. If NULL, the counts are grouped by 'fov' and 'codeclass'. Defaults to NULL.}

\item{fig_save}{A logical value indicating whether to save the plot as an image file. Defaults to FALSE.}

\item{fig_width}{A numeric value specifying the width of the saved plot image in inches. Defaults to 10.}

\item{fig_height}{A numeric value specifying the height of the saved plot image in inches. Defaults to 8.}

\item{out_dir}{A string specifying the path to the directory where the plot will be saved if \code{fig_save} is TRUE. If NULL and \code{fig_save} is TRUE, an error is raised.}
}
\value{
A ggplot object representing the bar plot of transcript counts by code class.
}
\description{
This function plots the counts of transcripts grouped by code class and optionally by another specified column from a raw transcript data file or data frame.
Optionally, the plot can be saved as an image file.
}
\details{
The function reads the raw transcript data from the specified CSV file or data frame. It then groups the data by code class and optionally by another specified column (\code{group_by}).
The resulting plot shows the proportion of transcripts for each code class within each group.

If \code{fig_save} is TRUE and \code{out_dir} is not provided, an error message is displayed. The plot is saved as a PNG file in the specified output directory if \code{fig_save} is TRUE.
}
\examples{
\dontrun{
plt <- PlotTranscriptByCodeClass("path/to/raw_transcripts.csv")
plt <- PlotTranscriptByCodeClass(raw_transcripts, group_by = "gene", fig_save = TRUE, out_dir = "path/to/output/directory")
}

}
