% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTranscriptCounts.R
\name{PlotTranscriptCounts}
\alias{PlotTranscriptCounts}
\title{Plot Transcript Counts}
\usage{
PlotTranscriptCounts(
  raw_transcripts_file,
  group_by = "fov",
  colour_by = NULL,
  fig_save = FALSE,
  fig_width = 10,
  fig_height = 8,
  out_dir = NULL
)
}
\arguments{
\item{raw_transcripts_file}{A character string representing the path to a CSV file containing raw transcript data, or a data frame with the transcript data.}

\item{group_by}{A string specifying the column name by which to group the transcript counts. Defaults to "fov".}

\item{colour_by}{A string specifying the column name to use for coloring the bars in the plot. If NULL, no coloring is applied. Defaults to NULL.}

\item{fig_save}{A logical value indicating whether to save the plot as an image file. Defaults to FALSE.}

\item{fig_width}{A numeric value specifying the width of the saved plot image in inches. Defaults to 10.}

\item{fig_height}{A numeric value specifying the height of the saved plot image in inches. Defaults to 8.}

\item{out_dir}{A string specifying the path to the directory where the plot will be saved if \code{fig_save} is TRUE. If NULL and \code{fig_save} is TRUE, an error is raised.}
}
\value{
A ggplot object representing the bar plot of transcript counts.
}
\description{
This function plots the counts of transcripts grouped by a specified column from a raw transcript data frame.
Optionally, the plot can be saved as an image file.
}
\details{
The function reads the raw transcript data from the specified CSV file or data frame. It then groups the data by the specified column (\code{group_by}) and optionally colors the bars by another specified column (\code{colour_by}).
The resulting plot shows the number of transcripts for each group, with a red dashed line indicating the median value.

If \code{fig_save} is TRUE and \code{out_dir} is not provided, an error message is displayed. The plot is saved as a PNG file in the specified output directory if \code{fig_save} is TRUE.
}
\examples{
\dontrun{
plt <- PlotTranscriptCounts("path/to/raw_transcripts.csv", group_by = "fov")
plt <- PlotTranscriptCounts(raw_transcripts, group_by = "fov", colour_by = "gene", fig_save = TRUE, out_dir = "path/to/output/directory")
}

}
