% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTranscriptLossProportion.R
\name{PlotTranscriptLossProportion}
\alias{PlotTranscriptLossProportion}
\title{Plot Transcript Loss Proportion}
\usage{
PlotTranscriptLossProportion(
  raw_transcripts,
  codeclass_filter = "Endogenous",
  keep_z_stacks = NULL,
  fig_save = FALSE,
  fig_width = 10,
  fig_height = 8,
  out_dir = NULL
)
}
\arguments{
\item{raw_transcripts}{A data frame containing raw transcript data.}

\item{codeclass_filter}{A string specifying the code class to filter transcripts by. Defaults to "Endogenous".}

\item{keep_z_stacks}{A vector of z-stacks to keep. If NULL, all z-stacks are kept. Defaults to NULL.}

\item{fig_save}{A logical value indicating whether to save the plot as an image file. Defaults to FALSE.}

\item{fig_width}{A numeric value specifying the width of the saved plot image in inches. Defaults to 10.}

\item{fig_height}{A numeric value specifying the height of the saved plot image in inches. Defaults to 8.}

\item{out_dir}{A string specifying the path to the directory where the plot will be saved if \code{fig_save} is TRUE. If NULL and \code{fig_save} is TRUE, an error is raised.}
}
\value{
A ggplot object representing the bar plot of transcript loss proportions.
}
\description{
This function visualizes the proportion of transcripts that are lost or retained across z-stacks for each field of view (FOV), filtered by code class.
Optionally, the plot can be saved as an image file.
}
\details{
The function filters the raw transcript data by the specified code class, calculates the number of transcripts per z-stack for each FOV, and determines which z-stacks to keep based on the provided \code{keep_z_stacks} parameter.
The resulting plot shows the proportion of transcripts in the selected z-stacks for each FOV.

If \code{fig_save} is TRUE and \code{out_dir} is not provided, an error message is displayed. The plot is saved as a PNG file in the specified output directory if \code{fig_save} is TRUE.
}
\examples{
\dontrun{
plt <- PlotTranscriptLossProportion(raw_transcripts)
plt <- PlotTranscriptLossProportion(raw_transcripts, codeclass_filter = "Endogenous", keep_z_stacks = c(1, 2, 3), fig_save = TRUE, out_dir = "path/to/output/directory")
}

}
