% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscripts.R
\name{plotTranscripts}
\alias{plotTranscripts}
\title{Plot transcript summary}
\usage{
plotTranscripts(transcripts, by = "fov", codeclass = FALSE)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{by}{a character vector specifying the column used to calculate and plot}

\item{codeclass}{a logical to determine if the code class distribution is to be plotted}
}
\value{
A ggplot object.
}
\description{
Initialises a ggplot object of either the number of raw transcripts or the code class distribution.
}
\examples{
\dontrun{

plotTranscripts(transcripts)
plotTranscripts(transcripts, by = "SampleID")
plotTranscripts(transcripts, by = "SampleID", codeclass = TRUE)
}
}
