% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscriptsByCellBoundary.R
\name{plotTranscriptsByCellBoundary}
\alias{plotTranscriptsByCellBoundary}
\title{Plot proportion of transcripts outside cell boundaries}
\usage{
plotTranscriptsByCellBoundary(transcripts, estimateChangepoint = FALSE)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{estimateChangepoint}{a logical to determine if a change point should be estimated}
}
\value{
A ggplot object.
}
\description{
Initialises a ggplot object to plot the proportion of transcripts detected outside cell boundaries by FOV and Z-slice. A change point in this proportion can be estimated using envcpt.
}
\examples{
\dontrun{

plotTranscriptsByCellBoundary(transcripts)
plotTranscriptsByCellBoundary(transcripts, estimateChangepoint = TRUE)
}
}
