% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscriptsPerCell.R
\name{plotTranscriptsPerCell}
\alias{plotTranscriptsPerCell}
\title{Plot relationship cell and transcript number}
\usage{
plotTranscriptsPerCell(transcripts, by = "fov", detectOutliers = FALSE)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{by}{a character vector specifying the column used to}

\item{detectOutliers}{a logical to determine if outliers should be detected}
}
\value{
A ggplot object.
}
\description{
Initialises a ggplot object of the relationship between the number of cells and the number of raw transcripts. Outliers can be detected by modelling this relationship using lmrob.
}
\examples{
\dontrun{

plotTranscriptsPerCell(transcripts)
plotTranscriptsPerCell(transcripts, detectOutliers = TRUE)
plotTranscriptsPerCell(transcripts, by = "SampleID", detectOutliers = TRUE)
}
}
