% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTranscriptsPerZslice.R
\name{plotTranscriptsPerZslice}
\alias{plotTranscriptsPerZslice}
\title{Create heatmap of transcript distribution across Z-slices}
\usage{
plotTranscriptsPerZslice(transcripts, by = "fov")
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{by}{a character vector specifying the column used to calculate and plot}
}
\value{
A pheatmap object.
}
\description{
A function to draw a heatmap to visualise the transcript distribution across Z-slices.
}
\examples{
\dontrun{

plotTranscriptsPerZslice(transcripts)
plotTranscriptsPerZslice(transcripts, by = "SampleID")
}
}
