% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareGEXMatrix.R
\name{PrepareGEXMatrix}
\alias{PrepareGEXMatrix}
\title{Prepare Gene Expression (GEX) Matrix}
\usage{
PrepareGEXMatrix(
  raw_transcripts,
  z_stacks,
  output_dir = NULL,
  save_matrix = FALSE
)
}
\arguments{
\item{raw_transcripts}{A data frame containing raw transcript data or a character string representing the path to a CSV file containing the data.}

\item{z_stacks}{A vector of z-stacks to keep in the analysis.}

\item{output_dir}{A string specifying the path to the directory where the GEX matrix will be saved if \code{save_matrix} is TRUE.}

\item{save_matrix}{A logical value indicating whether to save the GEX matrix as a CSV file. Defaults to FALSE.}
}
\value{
A data frame representing the cleaned GEX matrix.
}
\description{
This function prepares a GEX matrix from raw transcript data, filtering by specified z-stacks and removing system control probes and empty cells.
Optionally, the matrix can be saved as a CSV file.
}
\details{
The function reads the raw transcript data from the specified CSV file or data frame. It filters out cells with \code{CellId} equal to 0 and keeps only the specified z-stacks.
It also removes probes labeled as 'SystemControl' and creates a unique cell identifier. The resulting GEX matrix is then generated and optionally saved as a CSV file.

If \code{save_matrix} is TRUE and \code{output_dir} is not provided, an error message is displayed. The matrix is saved as a CSV file in the specified output directory if \code{save_matrix} is TRUE.
}
\examples{
\dontrun{
gex_matrix <- PrepareGEXMatrix(raw_transcripts, z_stacks = c(1, 2, 3), output_dir = "path/to/output/directory")
gex_matrix <- PrepareGEXMatrix(raw_transcripts, z_stacks = c(1, 2, 3), output_dir = "path/to/output/directory", save_matrix = FALSE)
}

}
