% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setQualityFlags.R
\name{setQualityFlags}
\alias{setQualityFlags}
\title{Add quality flags}
\usage{
setQualityFlags(transcripts)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}
}
\value{
A data frame containing raw transcripts with CellBoundaryFlag and SystemControlFlag columns.
}
\description{
Function that adds flags for system control probes and transcripts that are located outside cell boundaries. Required to detect transcript dispersion.
}
\examples{
\dontrun{

transcripts <- setQualityFlags(transcripts)
}
}
