% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setZsliceFlag.R
\name{setZsliceFlag}
\alias{setZsliceFlag}
\title{Add Z-slice flag}
\usage{
setZsliceFlag(transcripts, index)
}
\arguments{
\item{transcripts}{a data frame containing raw transcripts}

\item{index}{An integer that determines which Z-slices will be flagged}
}
\value{
A data frame containing raw transcripts with the ZstackFlag column.
}
\description{
Function to add a flag for Z-slices that are to be removed. Should only be used after having run plotTranscriptsByCellBoundary().
}
\examples{
\dontrun{

transcripts <- setZsliceFlag(transcripts, 7)
}
}
