% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateMetadataAndPolygons.R
\name{UpdateMetadataAndPolygons}
\alias{UpdateMetadataAndPolygons}
\title{Update Metadata and Polygons Data Frames}
\usage{
UpdateMetadataAndPolygons(
  clean_GEX,
  metadata,
  polygons = NULL,
  save_flat_files = FALSE,
  output_dir = NULL
)
}
\arguments{
\item{clean_GEX}{A data frame or a character string representing the file path to a CSV file containing the clean_GEX data. The data frame must have row names corresponding to cell identifiers.}

\item{metadata}{A data frame or a character string representing the file path to a CSV file containing the metadata. The data frame must have a column named 'cell'.}

\item{polygons}{A data frame or a character string representing the file path to a CSV file containing the polygons data (optional). The data frame must have a column named 'cell'.}

\item{save_flat_files}{A logical value indicating whether to save the updated metadata and polygons to flat files. Default is FALSE.}

\item{output_dir}{A character string representing the directory where the updated flat files will be saved. Required if save_flat_files is TRUE.}
}
\value{
A list containing the updated metadata and polygons data frames.
}
\description{
This function updates metadata and polygons data frames based on the provided clean_GEX data frame. It recalculates RNA counts and features, filters and arranges the metadata and polygons to match clean_GEX, and optionally saves the updated data frames to specified flat files.
}
\details{
The function performs the following steps:
1. Reads the clean_GEX data from the provided data frame or file path.
2. Reads the metadata from the provided data frame or file path.
3. Ensures the metadata contains a column named 'cell'.
4. Filters and arranges the metadata to match the row names of clean_GEX.
5. Identifies negative probes (columns containing 'Neg') and regular probes.
6. Recalculates nCount_RNA and nFeature_RNA excluding negative probes.
7. Recalculates nCount_negprobes and nFeature_negprobes.
8. Removes columns 'nCount_falsecode' and 'nFeature_falsecode' if they exist in the metadata.
9. Optionally reads the polygons data from the provided data frame or file path, filters and arranges it to match the row names of clean_GEX.
10. Optionally saves the updated metadata and polygons to the specified output directory.
}
\examples{
\dontrun{
# Example usage:
updated_data <- UpdateMetadataAndPolygons(clean_GEX = "path/to/clean_GEX.csv",
                                          metadata = "path/to/metadata.csv",
                                          polygons = "path/to/polygons.csv",
                                          save_flat_files = TRUE,
                                          output_dir = "path/to/output/dir")
}

}
