% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualizeZStackDistribution.R
\name{VisualizeZStackDistribution}
\alias{VisualizeZStackDistribution}
\title{Visualize Z-Stack Distribution}
\usage{
VisualizeZStackDistribution(
  raw_transcripts_file,
  binarised = FALSE,
  fig_save = FALSE,
  fig_width = 10,
  fig_height = 8,
  out_dir = NULL
)
}
\arguments{
\item{raw_transcripts_file}{A character string representing the path to a CSV file containing raw transcript data, or a data frame with the transcript data.}

\item{binarised}{A logical value indicating whether to binarize the data. If TRUE, the data is binarized to indicate presence (1) or absence (0) of transcripts. Defaults to FALSE.}

\item{fig_save}{A logical value indicating whether to save the plot as an image file. Defaults to FALSE.}

\item{fig_width}{A numeric value specifying the width of the saved plot image in inches. Defaults to 10.}

\item{fig_height}{A numeric value specifying the height of the saved plot image in inches. Defaults to 8.}

\item{out_dir}{A string specifying the path to the directory where the plot will be saved if \code{fig_save} is TRUE. If NULL and \code{fig_save} is TRUE, an error is raised.}
}
\value{
A pheatmap object representing the heatmap of transcript distribution across z-stacks for each FOV.
}
\description{
This function visualizes the distribution of transcripts across z-stacks for each field of view (FOV).
Optionally, the data can be binarized and the plot can be saved as an image file.
}
\details{
The function reads the raw transcript data from the specified CSV file or data frame. It then calculates the number of transcripts per z-stack for each FOV.
The resulting heatmap shows the distribution of transcripts, either scaled or binarized, across z-stacks for each FOV.

If \code{fig_save} is TRUE and \code{out_dir} is not provided, an error message is displayed. The plot is saved as a PNG file in the specified output directory if \code{fig_save} is TRUE.
}
\examples{
\dontrun{
plt <- VisualizeZStackDistribution("path/to/raw_transcripts.csv")
plt <- VisualizeZStackDistribution(raw_transcripts, binarised = TRUE, fig_save = TRUE, out_dir = "path/to/output/directory")
}

}
