% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZPlot.R
\name{ZPlot}
\alias{ZPlot}
\title{Plot Transcripts by Z-Stack}
\usage{
ZPlot(
  raw_transcripts,
  fovs,
  fig_save = FALSE,
  out_dir = NULL,
  fig_width = 8,
  fig_height = 6
)
}
\arguments{
\item{raw_transcripts}{A data frame containing raw transcript data or a character string representing the path to a CSV file containing the data.}

\item{fovs}{A vector of FOV identifiers for which the visualizations will be generated.}

\item{fig_save}{A logical value indicating whether to save the plots as image files. Defaults to FALSE.}

\item{out_dir}{A string specifying the path to the directory where the plots will be saved if \code{fig_save} is TRUE. If NULL and \code{fig_save} is TRUE, an error is raised.}

\item{fig_width}{A numeric value specifying the width of the saved plot images in inches. Defaults to 8.}

\item{fig_height}{A numeric value specifying the height of the saved plot images in inches. Defaults to 6.}
}
\value{
A list of ggplot objects representing the plots of transcript distributions for each specified FOV.
}
\description{
This function visualizes the distribution of transcripts across z-stacks for specified fields of view (FOVs).
Optionally, the plots can be saved as image files.
}
\details{
The function reads the raw transcript data from the specified CSV file or data frame. For each specified FOV, it plots the distribution of transcripts across z-stacks, showing their x and y coordinates and coloring them by code class.
The resulting plot shows the spatial distribution of transcripts in the specified FOVs, faceted by z-stack.

If \code{fig_save} is TRUE and \code{out_dir} is not provided, an error message is displayed. The plots are saved as PNG files in the specified output directory if \code{fig_save} is TRUE.
}
\examples{
\dontrun{
plots <- ZPlot(raw_transcripts, fovs = c(1:10))
plots <- ZPlot(raw_transcripts, fovs = c(1:10), fig_save = TRUE, out_dir = "path/to/output/directory")
}

}
